Fix TypeError at runtime when importing types
A class cannot be inherited from a TypedDict and non-TypedDict at the same time so this raised a TypeError at runtime. So GenericWebsocketEvent has been renamed to BaseWebsocketEvent and is no longer a generic type. The type field now defaults to EventTypeRecv since it was the most common use case