Add documentation for luster.types
This commit adds documentation for luster.types and also implements some unfortunate workarounds for specific problems related to documentation.

Firstly, Sphinx does not properly document Literal type aliases. It completely ignores it's docstring and just documents it as 'alias of Literal[...]' which is not the desired output. Unfortunately, there was no workaround for this that I could find so for now: All enumerations are documented manually in types.rst rather than using autodoc.

Secondly, all the enums are moved to separate module. This is done to avoid runtime imports of enums. Sphinx tends to document type aliases used in annotations as 'Literal[...]' rather than the alias itself so we need to avoid importing them at runtime so the annotation is the string at runtime.