Fix channel not being evicted during GroupLeave
This commit fixes group channel not being evicted when connected user leaves it. This is due to the fact that Revolt API does not send any delete event -- rather it sends a GroupLeave event.

This commit changes the connected user's data to be stored in state rather than client as this data is required by library for properly evicting channel from cache during GroupLeave event.